add dumped schemas
