+++ python/librar/mysql.py
@@ -3,8 +3,10 @@
 # Alternative license arrangements possible, contact me for more information
 """ Code for interacting with MySQL """
 
+import os
 import sys
 import json
+import yaml
 import inspect
 
 from MySQLdb import _mysql
@@ -16,7 +18,7 @@ from librar.log import log, debug, init as log_init
 
 ALL_CREDENTIALS = ["database", "username", "password", "server"]
 
-numbers = {"tinyint", "int", "decimal"}
+INTS = {"tinyint", "int", "decimal"}
 
 def sort_elm(fld_elm):
     return fld_elm["Field"]
@@ -42,6 +44,96 @@ my_conv[FIELD_TYPE.BLOB] = convert_string
 my_conv[FIELD_TYPE.TINY] = int
 my_conv[FIELD_TYPE.DECIMAL] = int
 my_conv[FIELD_TYPE.NEWDECIMAL] = int
+
+
+def add_join_items(new_schema):
+    """ add `join` items to columns that join """
+    jns = new_schema[":more:"]["joins"]
+    for table in list(new_schema):
+        if table[0] != ":":
+            for col in list(new_schema[table]["columns"]):
+                dst = find_join_dest(table, col, jns)
+                if dst is not None:
+                    tblcol = dst.split(".")
+                    new_schema[table]["columns"][col]["join"] = {"table": tblcol[0], "column": tblcol[1]}
+
+
+def find_join_dest(table, col, jns):
+    """ does this {rable.col} have join dest in {jns} """
+    long = table + "." + col
+    if long in jns and jns[long] != long:
+        return jns[long]
+    if col in jns and jns[col] != long:
+        return jns[col]
+    return None
+
+
+def load_more_schema(new_schema):
+    """ load users file of additional schema information """
+    new_schema[":more:"] = {}
+    filename = f"{os.environ['BASE']}/etc/pyrar.yml"
+    if os.path.isfile(filename):
+        with open(filename, encoding="utf-8") as file:
+            data = file.read()
+            new_schema[":more:"] = yaml.load(data, Loader=yaml.FullLoader)
+            return
+
+
+
+def schema_of_col(new_schema, col):
+    """ convert MySQL column description into JSON schema """
+    this_field = {}
+    this_type = col["Type"]
+    this_places = 0
+    if this_type.find(" unsigned") >= 0:
+        this_type = this_type.split()[0]
+        this_field["unsigned"] = True
+
+    pos = this_type.find("(")
+    if pos >= 0:
+        this_size = this_type[pos + 1:-1]
+        this_type = this_type[:pos]
+        if this_size.find(",") >= 0:
+            tmp = this_size.split(",")
+            this_field["size"] = int(tmp[0])
+            this_field["places"] = int(tmp[1])
+            this_places = int(tmp[1])
+        else:
+            if ((int(this_size) == 1 and this_type == "tinyint")
+                    or (":more:" in new_schema and "is_boolean" in new_schema[":more:"]
+                        and col["Field"] in new_schema[":more:"]["is_boolean"])):
+                this_type = "boolean"
+                if "unsigned" in this_field:
+                    del this_field["unsigned"]
+            else:
+                this_field["size"] = int(this_size)
+
+    this_field["type"] = this_type
+    if col["Extra"] == "auto_increment":
+        this_field["serial"] = True
+
+    this_field["null"] = (col["Null"] == "YES")
+    plain_int = test_plain_int(this_type, this_places)
+    this_field["is_plain_int"] = plain_int
+    if col["Default"] is not None:
+        defval = col["Default"]
+        if plain_int:
+            defval = int(defval)
+        elif this_type == "boolean":
+            defval = (int(defval) == 1)
+        this_field["default"] = defval
+
+    return this_field
+
+
+def test_plain_int(this_type, this_places):
+    """ return True if {this_type} with {this_places}
+        # of decimal places is in INT """
+    if this_type in INTS:
+        return True
+    if this_type == "decimal" and this_places == 0:
+        return True
+    return False
 
 
 def event_log(other_items, stack_pos=2):
@@ -164,6 +256,15 @@ class MariaDB:
         self.tcp_cnx.commit()
         return affected_rows, lastrowid
 
+    def get_cols(self,table):
+        if self.schema is None:
+            self.make_schema()
+            if self.schema is None:
+                return None
+        if table in self.schema and "columns" in self.schema[table]:
+            return self.schema[table]["columns"]
+        return None
+
     def sql_close(self):
         self.tcp_cnx.close()
 
@@ -179,10 +280,9 @@ class MariaDB:
         return ok is not None
 
     def sql_insert(self, table, column_vals, ignore=False):
-        if table in static.AUTO_CREATED_AMENDED_DT:
-            for col in ["amended_dt", "created_dt"]:
-                if col not in column_vals:
-                    column_vals[col] = None
+        if (cols := self.get_cols(table)) is not None:
+            for col in static.NOW_DATE_FIELDS:
+                column_vals[col] = None
         with_ignore = "ignore" if ignore else ""
         return self.sql_exec(f"insert {with_ignore} into {table} set " + data_set(column_vals, ",", is_set=True))
 
@@ -192,7 +292,7 @@ class MariaDB:
         return (ret is not None) and (self.tcp_cnx.affected_rows() > 0)
 
     def sql_update_one(self, table, column_vals, where):
-        if table in static.AUTO_CREATED_AMENDED_DT and isinstance(column_vals, dict):
+        if (cols := self.get_cols(table)) is not None and "amended_dt" in cols and isinstance(column_vals, dict):
             column_vals["amended_dt"] = None
         return self.sql_update(table, column_vals, where, 1)
 
@@ -317,6 +417,21 @@ class MariaDB:
 
         return self.tcp_cnx is not None
 
+    def add_indexes_to_schema(self, new_schema, table):
+        """ Add index info for {table} to {new_schema} """
+        ok, reply = self.run_select("show index from " + table)
+        if not ok:
+            raise ValueError(f"Could not get indexes for '{table}'")
+
+        new_schema[table]["indexes"] = {}
+        for col in reply:
+            key = col["Key_name"] if col["Key_name"] != "PRIMARY" else ":primary:"
+            if key not in new_schema[table]["indexes"]:
+                new_schema[table]["indexes"][key] = {}
+                new_schema[table]["indexes"][key]["columns"] = []
+            new_schema[table]["indexes"][key]["columns"].append(col["Column_name"])
+            new_schema[table]["indexes"][key]["unique"] = col["Non_unique"] == 0
+
     def make_schema(self):
         schema = {}
         ok, ret = self.run_select("show tables")
@@ -328,40 +443,13 @@ class MariaDB:
             this_tbl["columns"] = {}
             cols = list(ret)
             cols.sort(key=sort_elm)
-            for each_col in cols:
-                this_tbl["columns"][each_col["Field"]] = {}
-                fld_type = each_col["Type"]
-                if fld_type.find(" unsigned") >= 0:
-                    fld_type = fld_type.split()[0]
-                    this_tbl["columns"][each_col["Field"]]["unsigned"] = True
+            for col in cols:
+                schema[table]["columns"][col["Field"]] = schema_of_col(schema, col)
+            self.add_indexes_to_schema(schema, table)
 
-                pos = fld_type.find("(")
-                if pos >= 0:
-                    fld_sz = fld_type[pos + 1:-1]
-                    if fld_sz[-2:] == ",0":
-                        fld_sz = fld_sz[:-2]
-                    this_tbl["columns"][each_col["Field"]]["size"] = int(fld_sz)
-                    fld_type = fld_type[:pos]
-
-                this_tbl["columns"][each_col["Field"]]["type"] = fld_type
-                if each_col["Extra"] == "auto_increment":
-                    this_tbl["columns"][each_col["Field"]]["serial"] = True
-
-                this_tbl["columns"][each_col["Field"]]["null"] = each_col["Null"] == "YES"
-                if each_col["Default"] is not None:
-                    if fld_type in numbers:
-                        this_tbl["columns"][each_col["Field"]]["default"] = int(each_col["Default"])
-                    else:
-                        this_tbl["columns"][each_col["Field"]]["default"] = each_col["Default"]
-
-            ok, ret = self.run_select("show index from " + table)
-            this_tbl["indexes"] = {}
-            for each_idx in ret:
-                if each_idx["Key_name"] not in this_tbl["indexes"]:
-                    this_tbl["indexes"][each_idx["Key_name"]] = {}
-                    this_tbl["indexes"][each_idx["Key_name"]]["columns"] = []
-                this_tbl["indexes"][each_idx["Key_name"]]["columns"].append(each_idx["Column_name"])
-                this_tbl["indexes"][each_idx["Key_name"]]["unique"] = each_idx["Non_unique"] == 0
+        load_more_schema(schema)
+        if ":more:" in schema and "joins" in schema[":more:"]:
+            add_join_items(schema)
 
         self.schema = schema
 
@@ -372,7 +460,7 @@ sql_server = MariaDB()
 
 def main():
     log_init(with_debug=True)
-    sql_server.connect("webui")
+    sql_server.connect("admin")
     print(json.dumps(sql_server.schema,indent=3))
     sys.exit(0)
