+++ python/librar/static.py
@@ -51,6 +51,3 @@ POW10 = [
 ]
 
 NOW_DATE_FIELDS = ["when_dt", "amended_dt", "created_dt", "deleted_dt"]
-AUTO_CREATED_AMENDED_DT = [
-    "paymwnts", "domains", "backend", "order_items", "orders", "sales_items", "session_keys", "users"
-]
